+++ examples/expoShowcase/constants/index.ts
@@ -1,0 +1,2 @@
+export const MAX_FONT_SCALE = 100;
+export const DISABLE_FONT_SCALE = false;