+++ examples/expoShowcase/router/index.tsx
@@ -1,0 +1,40 @@
+import 'react-native-gesture-handler';
+import React, { ReactNode } from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createDrawerNavigator } from '@react-navigation/drawer';
+import { View } from 'react-native';
+import { NavDrawerProps, NavigationDrawer } from './NavigationDrawer';
+import { App } from '../Home';
+import { createStackNavigator } from '@react-navigation/stack';
+
+const Drawer = createDrawerNavigator();
+
+export type RootStackParamList = {
+    App: undefined;
+    NavigationDrawer: undefined;
+};
+
+const RootStack = createStackNavigator<RootStackParamList>();
+
+const CustomDrawerContent = (props: any): any => (
+    <View style={{ height: '100%' }}>
+        <NavigationDrawer {...props} />
+    </View>
+);
+
+export const MainRouter = (): any => (
+    <NavigationContainer>
+        <Drawer.Navigator
+            initialRouteName="App"
+            drawerStyle={{ backgroundColor: 'transparent', width: '80%' }}
+            screenOptions={{
+                headerShown: false,
+                drawerType: 'front',
+                drawerStyle: { backgroundColor: 'transparent' },
+            }}
+            drawerContent={(props: NavDrawerProps): ReactNode => <CustomDrawerContent {...props} />}
+        >
+            <RootStack.Screen name="App" component={App} />
+        </Drawer.Navigator>
+    </NavigationContainer>
+);