+++ packages/auth-workflows/jest.config.cjs
@@ -16,9 +16,9 @@ module.exports = {
         "./node_modules/react-native-gesture-handler/jestSetup.js",
         "./jestSetupFile.js"
     ],
-    "setupFilesAfterEnv": [
-        "<rootDir>setup-tests.js"
-    ],
+    // "setupFilesAfterEnv": [
+    //     "<rootDir>setup-tests.js"
+    // ],
     "moduleNameMapper": {
         ".+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$": "identity-obj-proxy"
     },