+++ packages/cli-templates/authentication-typescript/template/src/components/UserMenuComponent.tsx
@@ -81,7 +81,7 @@ export const UserMenuComponent: React.FC<UserMenuExampleProps> = (props) => {
             rightComponent: (
                 <SelectDropdown
                     defaultValue={languageOptions.find((option) => option.value === i18n.language)}
-                    onSelect={(item: any) => handleLanguageChange(item.value)}
+                    onSelect={(item: any) => void handleLanguageChange(item.value)}
                     data={languageOptions}
                     buttonStyle={{ backgroundColor: theme.colors.background }}
                     buttonTextStyle={{ color: theme.colors.primary }}