+++ scripts/linkProgressIcons.sh
@@ -18,6 +18,11 @@ mkdir -p "./examples/showcase/node_modules/@brightlayer-ui/react-native-progress
 rm -rf "./examples/workflow/node_modules/@brightlayer-ui/react-native-progress-icons"
 mkdir -p "./examples/workflow/node_modules/@brightlayer-ui/react-native-progress-icons/dist"
 
+rm -rf "./examples/expoShowcase/node_modules/@brightlayer-ui/react-native-progress-icons"
+mkdir -p "./examples/expoShowcase/node_modules/@brightlayer-ui/react-native-progress-icons/dist"
+
+rm -rf "./examples/workflowexample-expo/node_modules/@brightlayer-ui/react-native-progress-icons"
+mkdir -p "./examples/workflowexample-expo/node_modules/@brightlayer-ui/react-native-progress-icons/dist"
 echo -e "${GREEN}Done${NC}"
 
 echo -en "${BLUE}Copying build output into node_modules...${NC}";
@@ -26,6 +31,12 @@ cp ./packages/progress-icons/package.json ./examples/showcase/node_modules/@brig
 cp -r ./packages/progress-icons/dist/ ./examples/showcase/node_modules/@brightlayer-ui/react-native-progress-icons/dist
 cp ./packages/progress-icons/package.json ./examples/workflow/node_modules/@brightlayer-ui/react-native-progress-icons/package.json
 cp -r ./packages/progress-icons/dist/ ./examples/workflow/node_modules/@brightlayer-ui/react-native-progress-icons/dist
+
+cp ./packages/progress-icons/package.json ./examples/expoShowcase/node_modules/@brightlayer-ui/react-native-progress-icons/package.json
+cp -r ./packages/progress-icons/dist/ ./examples/expoShowcase/node_modules/@brightlayer-ui/react-native-progress-icons/dist
+
+cp ./packages/progress-icons/package.json ./examples/workflowexample-expo/node_modules/@brightlayer-ui/react-native-progress-icons/package.json
+cp -r ./packages/progress-icons/dist/ ./examples/workflowexample-expo/node_modules/@brightlayer-ui/react-native-progress-icons/dist
 echo -e "${GREEN}Done${NC}"
 
 echo -en "\r\n${BBLUE}Linking Components: ${NC}"
@@ -33,13 +44,29 @@ if [ ! -f ./examples/showcase/node_modules/@brightlayer-ui/react-native-progress
 if [ ! -s ./examples/showcase/node_modules/@brightlayer-ui/react-native-progress-icons/dist ];
     then
         if [ ! -f ./examples/showcase/node_modules/@brightlayer-ui/react-native-progress-icons/dist/index.js ];
-        then echo -e "${BRED}Progress Icons Not Linked${NC}" && exit 1;
+        then echo -e "${BRED}Progress Icons Not Linked to showcase${NC}" && exit 1;
         fi;
 fi
 if [ ! -f ./examples/workflow/node_modules/@brightlayer-ui/react-native-progress-icons/package.json ]; then echo -e "${BRED}Progress Icons Not Linked${NC}" && exit 1; fi
 if [ ! -s ./examples/workflow/node_modules/@brightlayer-ui/react-native-progress-icons/dist ];
     then
         if [ ! -f ./examples/workflow/node_modules/@brightlayer-ui/react-native-progress-icons/dist/index.js ];
+        then echo -e "${BRED}Progress Icons Not Linked to workflow${NC}" && exit 1;
+        fi;
+fi
+
+if [ ! -f ./examples/expoShowcase/node_modules/@brightlayer-ui/react-native-progress-icons/package.json ]; then echo -e "${BRED}Progress Icons Not Linked${NC}" && exit 1; fi
+if [ ! -s ./examples/expoShowcase/node_modules/@brightlayer-ui/react-native-progress-icons/dist ];
+    then
+        if [ ! -f ./examples/expoShowcase/node_modules/@brightlayer-ui/react-native-progress-icons/dist/index.js ];
+        then echo -e "${BRED}Progress Icons Not Linked to expo showcase${NC}" && exit 1;
+        fi;
+fi
+
+if [ ! -f ./examples/workflowexample-expo/node_modules/@brightlayer-ui/react-native-progress-icons/package.json ]; then echo -e "${BRED}Progress Icons Not Linked${NC}" && exit 1; fi
+if [ ! -s ./examples/workflowexample-expo/node_modules/@brightlayer-ui/react-native-progress-icons/dist ];
+    then
+        if [ ! -f ./examples/workflowexample-expo/node_modules/@brightlayer-ui/react-native-progress-icons/dist/index.js ];
         then echo -e "${BRED}Progress Icons Not Linked${NC}" && exit 1;
         fi;
 fi