+++ packages/cli-templates/blank-typescript/CHANGELOG.md
@@ -1,23 +1,32 @@
 # Changelog
 
+## v4.0.0 (Unreleased)
+
+### Changed
+
+- Update to build with React Native 0.80.
+- Upgraded the BLUI packages to use latest version.
+- Package dependencies upgraded to use latest version.
+- Changed cli to use latest template command of RN Community.
+
 ## v3.0.0 (June 5, 2024)
 
 ### Changed
 
--   Update to build with React Native 0.73.
--   Updated the templates to use React Native CLI
+- Update to build with React Native 0.73.
+- Updated the templates to use React Native CLI
 
 ## v2.1.0 (November 1, 2022)
 
 ### Changed
 
--   Update to build with React Native 0.70.
+- Update to build with React Native 0.70.
 
 ## v2.0.0 (December 16, 2021)
 
 ### Changed
 
--   Changed package namespace from `@pxblue` to `@brightlayer-ui`.
+- Changed package namespace from `@pxblue` to `@brightlayer-ui`.
 
 ## Package Migration Notice
 
@@ -40,22 +49,27 @@ Previous versions listed after this indicator refer to our deprecated `@pxblue` 
 ## v1.3.0 (September 30, 2021)
 
 ### Changed
+
 - Upgrade PX Blue packages
 
 ## v1.2.0 (July 26, 2021)
 
 ### Changed
+
 - Updated version of react-native-modal dependency to 12.0.2
 
 ## v1.1.0 (March 31, 2021)
 
 ### Changed
--   Updated dependencies to use latest packages.
+
+- Updated dependencies to use latest packages.
 
 ### Fixed
--   Jest configuration allows default tests to pass.
+
+- Jest configuration allows default tests to pass.
 
 ## v1.0.0 (March 1, 2021)
 
 ### Added
--   Initial blank template for TypeScript projects
+
+- Initial blank template for TypeScript projects